metodos DAO_SQL modificados

+++ src/DAL/MySQL/scripts/createTables/tableRoomPhotos.sql
@@ -1,6 +1,6 @@
 use mirandahoteldb;
 CREATE TABLE room_photos(
     id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-    roomId TINYINT NOT NULL,
+    roomId INT NOT NULL,
     url TEXT NOT NULL
 );
